Split Relation methods into Collection, Mutation, and SerializationMethods
Move some Relation tests in the direction of their implementations as well.

This sets us up for a pending NullRelation refactor.
